backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.44
psycopg[binary]>=3.1,<4.0
pydantic==2.11.9
pydantic-settings==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
pypdf==5.9.0
python-docx==1.2.0
numpy==2.3.2
google-generativeai==0.8.5
httpx==0.28.1
pytest==8.4.2
pytest-asyncio==1.1.0
